+++ CHANGELOG.md
@@ -11,6 +11,7 @@
         - Removed `delay` option from the constructor
         - Removed `appearanceDelay` property
         - Removed adding `.init` class to the main container at the first stage change
+- Changed `createLoadDataState()` to take a dataset factory function instead of a load data request, enhancing its reusability
 - Fixed value counting in the `signature` view when a property contains duplicated objects, which previously led to incorrect `optional` label
 - Fixed warning "Added non-passive event listener..." in Chrome caused by CodeMirror (https://github.com/codemirror/codemirror5/issues/6735)
 - Fixed missed error on progressbar's value when the progressbar finishes with error